BaseAdapter changed from interface to SafeArgs. Adding fragments and Vmodels.
Changed adapters to use a combination of binding and safeargs for navigation. SafeArg value is passed to the next fragment in navigation hierarchy. Argument is then passed to ViewModel and finally repository for retrieval of data.

Also set-up simple ItemListFragment and populated recycler list to test firestore queries. Planning to change data structure. Currently there are 3 levels of normalisation and it's far too complicated.

Planning to reduce data reads by using pagination.