Added AddItemRepo fuctionality plus ImageUploadManager
ImageUploadManager added to assist in the upload of item images to firebase storage. Images are rescaled and compressed before uploading. Added appropriate error handling through a combination of simple interface observers and SingleEventLiveData (where life cycle observation is neccessary - ie - in the view)

Future plan is to use ImageUploadManager as a singleton to handle multiple uploads, offline processing, etc. For now - user must be online to upload images.