Edited BindingAdapters. Object passed to ItemFrag
Created an static ItemPile bus, so that ItemPile objects may be passed from a selected adapter item (via binding view holder) and posted to the view model. ItemViewModel can then access this data and set it up as required for the View. This prevents having to fetch data a second time from Firestore.

The bus can also be used to pass the data to the item manager where it can be edited and uploaded to Firestore.